synth_sod/data_generation/flux_finetune/requirements.txt
torch>=2.4.0
torchvision>=0.19.0
numpy>=1.26.0
diffusers>=0.30.0
transformers>=4.45.0
accelerate>=1.0.0
peft>=0.13.0
huggingface-hub>=0.26.0
sentencepiece>=0.2.0
hydra-core>=1.3.0
omegaconf>=2.3.0
Pillow>=10.0.0
bitsandbytes>=0.44.0
prodigyopt>=1.0
wandb>=0.17.0
tqdm>=4.66.0
fire>=0.7.0
pandas>=2.0.0
